deepfake_model/requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
peft>=0.7.0

# Face Detection
insightface>=0.7.3
onnxruntime-gpu>=1.16.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data & Utils
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
tqdm>=4.65.0

# Training
wandb>=0.16.0

# Inference API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
